deep-spectral-segmentation/requirements.txt
accelerate
fire
opencv-python-headless
pillow
scikit-image
scipy
torch
torchvision
tqdm
pymatting
wandb
# polyaxon_client==0.6.1
omegaconf
hydra-core
SimpleCRF
scikit-learn
# pip install markupsafe==2.0.1 --force